chore: answer dto 분리
