+++ be/src/main/java/kr/codesquad/jazzmeet/inquiry/dto/request/answer/InquiryAnswerSaveRequest.java
@@ -1,9 +1,11 @@
-package kr.codesquad.jazzmeet.inquiry.dto.request;
+package kr.codesquad.jazzmeet.inquiry.dto.request.answer;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record InquiryAnswerUpdateRequest(
+public record InquiryAnswerSaveRequest(
+	@NotNull
+	Long inquiryId,
 	@NotNull
 	@Size(min = 1, max = 1000)
 	String content
