+++ be/src/main/java/kr/codesquad/jazzmeet/inquiry/dto/request/answer/InquiryAnswerUpdateRequest.java
@@ -1,11 +1,9 @@
-package kr.codesquad.jazzmeet.inquiry.dto.request;
+package kr.codesquad.jazzmeet.inquiry.dto.request.answer;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record InquiryAnswerSaveRequest(
-	@NotNull
-	Long inquiryId,
+public record InquiryAnswerUpdateRequest(
 	@NotNull
 	@Size(min = 1, max = 1000)
 	String content
