+++ be/src/main/java/kr/codesquad/jazzmeet/inquiry/dto/response/InquiryDetailResponse.java
@@ -1,5 +1,7 @@
 package kr.codesquad.jazzmeet.inquiry.dto.response;
 
+import kr.codesquad.jazzmeet.inquiry.dto.response.answer.InquiryAnswerDetail;
+
 public record InquiryDetailResponse(
 	Long id,
 	String content,
