+++ be/src/main/java/kr/codesquad/jazzmeet/inquiry/dto/response/answer/InquiryAnswerSaveResponse.java
@@ -1,4 +1,4 @@
-package kr.codesquad.jazzmeet.inquiry.dto.response;
+package kr.codesquad.jazzmeet.inquiry.dto.response.answer;
 
 import java.time.LocalDateTime;
